pyproject.toml
[tool.poetry]
name = "picpocket"
version = "0.1.0"
description = ""
authors = ["Patrick Buzzo <pbuzzo04@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.8"
pillow = "^7.2.0"
GPSPhoto = "^2.2.3"
exifread = "^2.1.2"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
